+++ twilight/src/main/java/com/siberanka/twilight/compiler/ResolvedJavaModel.java
@@ -5,6 +5,7 @@ import com.google.gson.JsonObject;
 
 import java.util.Map;
 
-record ResolvedJavaModel(String identifier, JsonArray elements, Map<String, String> textures, JsonObject display) {
+record ResolvedJavaModel(String identifier, JsonArray elements, Map<String, String> textures, JsonObject display,
+                         boolean handheld) {
     boolean isThreeDimensional() { return elements != null && !elements.isEmpty(); }
 }